Bullet Restructure, Removing Basic & Ice, Tweaks
Restructured bullet scripts. Bullet.cs is now attached to each bullet prefab. The elemental bullet scripts do not inherit from Bullet.cs, they are monobehaviour and are referenced to deal damage and inflict status effects as needed. Tower.cs now controls the bullet damage and speed, and contains all attributes related to status effects.

Removed the legacy Fortress Skill Tree and anything related to the Basic Tower.

Removed the ice tower and anything related to it. Finished implementing wind in its place.

Most number variable types have been changed from double/int to float.

+++ Assets/Scripts/Waves/Enemy.cs
@@ -60,7 +60,7 @@ public class Enemy : MonoBehaviour
 
         while (elapsedTime < duration)
         {
-            TakeDamage((int)burnDamage);
+            TakeDamage(burnDamage);
             elapsedTime += 1f; // Apply burn damage every second
             yield return new WaitForSeconds(1f);
         }
@@ -84,14 +84,14 @@ public class Enemy : MonoBehaviour
         isStunned = false;
     }
     
-    public void TakeDamage(int amount, string towerType = "Neutral") {
+    public void TakeDamage(float amount, string towerType = "Neutral") {
         int finalDamage = CalculateDamage(amount, towerType);
         health -= finalDamage;
 
         healthBar.fillAmount = (float)health / maxHealth;
     }
 
-    private int CalculateDamage(int baseDamage, string towerType)
+    private int CalculateDamage(float baseDamage, string towerType)
     {
         float multiplier = 1f;
         // EDIT THIS SWITCH STATEMENT TO ADD MORE ENEMY TYPES
